pyproject.toml
[project]
name = "highres_ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cartopy>=0.25.0",
    "catboost>=1.2.10",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "munch>=4.0.0",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "xarray>=2026.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
highres-ta = "highres_ta:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
